change normalization for calculation with subset of BZ
Drop \prod_i (k_{stop, i} - k_{start, i}) factor from tetra_volume.
With this change, the weights are the unitless quantity

\tilde{w}_{nj}^S = 1/V_S \int_S d^3k w_j(k) f_{FD}(\epsilon_n(k) - E_F)

Here S is the subset of the Brillouin zone included in the region delimited
by k_stop and k_start. When S is the full Brillouin zone, this is the
weight per unit cell. When S is a subset of the full Brillouin zone, the
total weight per volume \bar{w}_{nj} is given by

\bar{w}_{nj} = 1/(2\pi)^d \sum_S V_S \tilde{w}_{nj}^S

where d is the dimension of the system and V_S is the volume of the Brillouin
zone subset S.